vinyl: create empty run files
The planned design of single memory level requires us to store run's lsn
in the metadata log for all runs so that we can figure out which parts
of the zero level has been compacted for each particular range.
Currently, we discard empty runs right away: we don't create files for
them, we don't have them on the list, and so we can't write any info
about them to the metadata log. For the sake of single memory level,
this patch removes this little optimization: now empty runs are handled
just like normal runs.