vinyl: fix invalid dump of pinned mem
In case of range dump or compaction, we first add all in-memory trees to
the write iterator and remember dump_lsn = xm->lsn, and only then wait
for in-memory trees to be unpinned. On task completion we use dump_lsn
to filter dumped trees (they will have min_lsn <= dump_lsn, see
vy_range_dump_mems). If there's a pinned empty in memory tree on the
frozen list when we add it to the write iterator, its min_lsn will
become greater than dump_lsn (xm->lsn before wait) after we are done
waiting for it to be unpinned, and therefore we won't delete it on task
completion although we will have dumped it. This will break the
assumption made by the merge iterator that the same key doesn't have
duplicates, which will result in incorrect read in case of upserts.

To avoid this, make dump/compaction tasks wait for frozen in-memory
trees to be unpinned before adding them to the write iterator.

This fixes commit 8116f2096fef ("vinyl: pin in-memory trees on tx
prepare").

Closes #2318